+++ src/main/java/nl/bd/eindopdrachtjava/services/UserService.java
@@ -1,35 +1,42 @@
 package nl.bd.eindopdrachtjava.services;
 
+import lombok.AllArgsConstructor;
+import nl.bd.eindopdrachtjava.exceptions.ResourceAlreadyExistsException;
 import nl.bd.eindopdrachtjava.exceptions.ResourceNotFoundException;
 import nl.bd.eindopdrachtjava.models.entities.User;
 import nl.bd.eindopdrachtjava.models.requests.UserRegistrationRequest;
 import nl.bd.eindopdrachtjava.repositories.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
+@AllArgsConstructor
 @Service
 public class UserService implements UserDetailsService {
 
-    @Autowired
-    private UserRepository userRepository;
+    private final UserRepository userRepository;
+    private final BCryptPasswordEncoder bCryptPasswordEncoder;
 
     /**
      * Loads user from the database if username exists.
      */
     @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        UserDetails userDetails = (UserDetails) userRepository.findByUsername(username)
+    public User loadUserByUsername(String username) throws UsernameNotFoundException {
+        return userRepository.findByUsername(username)
                 .orElseThrow(() -> new UsernameNotFoundException("User does not exist!"));
-        return userDetails;
     }
 
     /**
      * Registers new user in the database.
      */
-    public User registerUser(UserRegistrationRequest userRegistrationRequest){
+    public User registerUser(UserRegistrationRequest userRegistrationRequest) throws ResourceAlreadyExistsException {
+        if (userExists(userRegistrationRequest)) {
+            throw new ResourceAlreadyExistsException(
+                    "User with username: " +
+                            userRegistrationRequest.getUsername() +
+                            " already exists!");
+        }
         User user = createUser(userRegistrationRequest);
         return userRepository.save(user);
     }
@@ -39,15 +46,27 @@ public class UserService implements UserDetailsService {
      */
     public User updateUser(UserRegistrationRequest userRegistrationRequest, Long userId) throws ResourceNotFoundException {
         return userRepository.findById(userId).map(user -> updatedUser(userRegistrationRequest, user))
-                .orElseThrow(() -> new ResourceNotFoundException("User with id " + userId + " was not found" ));
-
+                .orElseThrow(() -> new ResourceNotFoundException("User with id " + userId + " was not found"));
     }
 
     /**
      * Deletes user with given Id from database.
      */
-    public void deleteUser(Long userId) {
+    public void deleteUser(Long userId) throws ResourceNotFoundException {
+        if (userRepository.findById(userId).isEmpty()) {
+            throw new ResourceNotFoundException(
+                    "User with id " +
+                            userId +
+                            " was not found");
+        }
         userRepository.deleteById(userId);
+    }
+
+    /**
+     * Checks if user with given username exists
+     */
+    private boolean userExists(UserRegistrationRequest userRegistrationRequest) {
+        return userRepository.findByUsername(userRegistrationRequest.getUsername()).isPresent();
     }
 
     /**
@@ -56,7 +75,7 @@ public class UserService implements UserDetailsService {
     private User createUser(UserRegistrationRequest userRegistrationRequest) {
         return User.builder()
                 .username(userRegistrationRequest.getUsername())
-                .password(userRegistrationRequest.getPassword())
+                .password(bCryptPasswordEncoder.encode(userRegistrationRequest.getPassword()))
                 .role(userRegistrationRequest.getRole())
                 .build();
     }
@@ -65,23 +84,15 @@ public class UserService implements UserDetailsService {
      * Checks if userRegistrationRequest has empty fields, and uses the old values where no new values were given.
      */
     private User updatedUser(UserRegistrationRequest userRegistrationRequest, User user) {
-        Long tempUserId = user.getUserId();
-
-        if (userRegistrationRequest.getUsername() == null) {
-            user.setUsername(userRepository.findById(tempUserId).get().getUsername());
-        } else {
+        if (userRegistrationRequest.getUsername() != null) {
             user.setUsername(userRegistrationRequest.getUsername());
         }
 
-        if (userRegistrationRequest.getPassword() == null) {
-            user.setPassword(userRepository.findById(tempUserId).get().getPassword());
-        } else {
-            user.setPassword(userRegistrationRequest.getPassword());
+        if (userRegistrationRequest.getPassword() != null) {
+            user.setPassword(bCryptPasswordEncoder.encode(userRegistrationRequest.getPassword()));
         }
 
-        if (userRegistrationRequest.getRole() == null) {
-            user.setRole(userRepository.findById(tempUserId).get().getRole());
-        } else {
+        if (userRegistrationRequest.getRole() != null) {
             user.setRole(userRegistrationRequest.getRole());
         }
